jwks-server-py/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.30.6
pyjwt==2.9.0
cryptography==43.0.1
